[UserTimingL3] Checking against malformed performance.measure options
In the User Timing L3 spec
(https://w3c.github.io/user-timing/#measure-method), the
performance.measure function can take an options dictionary as its
second argument. If neither of the 'start' nor 'end' members is present
in the given options dictionary, a TypeError must be thrown.

This change adds checks for this case and raises the appropriate errors.
WPT coverage is also extended to check that user agents raise TypeErrors
as specified.

Change-Id: If47cf5803886cafb9dfab79c2cee1b6b572dbd3b