[css-text] Hang sequence of white spaces that may otherwise overflow
When a item overflows, we try to find a safe position that breaks the
item to avoid it. However, we should consider that sequence of white
space characters might have to be hanged, honoring the white-space
property.

We manage that case correctly the overflowing text is part of a
previously broken (kSuccess result from BreakText) item. However, if
there is an item that starts with a white space sequence and it
overflows, we always try to break it instead of hanging.

This patch address the issue by detecting such overflowing item,
verifying that it has a white space sequence and the white-space
property's value allows hanging it. In that case, we should rewind
the iterator, so that the further item_results are removed and skip
the rest of the overflowing handling logic.

Bug: 973358
Change-Id: I0e2ad5c5740f4ce592dd83b8ff3d224dbb45117e