[PW] Fix paint worklet canvas clip bound calculation
Right now when we initialize the paint canvas for paint worklet with a
scale factor or browser zoom, we do not record the zoom factor in the
canvas transform. As a result, when computing the clip bound we will
apply the wrong transformation matrix and thus sometimes the contents
can be completely clipped out while it should not.

This CL fixes the problem and adds a layout test.

Bug: 970783
Change-Id: Ie3709ea5845ca2b405245d4b36f9fe8f99052bad