Recalculate descendant styles when display changes to custom layout
Previously this only happened if the old display was custom layout,
but not when changing from a standard display to a custom one.
This was a typo, the same condition was checked twice for old_style,
but the second occurrence should check new_style instead.
The change matters when there are inline children that should be
blockified by the custom layout, without the patch they would remain
inline.

BUG=978990
TEST=external/wpt/css/css-layout-api/layout-child-inlines-dynamic.https.html

Change-Id: I73bcf679209dd4bc4a0d7a49f4345ee699bbf37e