Cleanup types in vr_service.mojom
This changes various vectors in the vr_service mojom to use the correct
gfx.mojom.geometry type.  It's largely mechanical, but the change from
array<float, 16> to gfx::Transform did introduce the need to transpose
the transforms being passed across that boundary.

For WMR, this resulted in a consistent way to convert from Matrix4x4 to
gfx::Transform, so some manual usages were consolidated to use a helper
function.

Bug:732403
Change-Id: I1eefaa90c7b7d5ea6fa03c01522422bd37f7290a